R1-Rev1: replace SY7200A LED driver with AP2502 current limiter
The SY7200A is a boost converter designed for series LED
strings, which require a high drive voltage. The external
circuitry allows some input power to pass through to the
output -- which is normally not a problem for LED strings
because this low voltage is not enough to overcome their
combined voltage drop.

However, the LCD uses 4 LEDs in parallel, and this low
voltage is enough to light the LEDs while the converter
is disabled, so the backlight can't be turned off.

Fix this by using the AP2502 4-channel current limiter
instead, which is basically a low side switch. LEDs are
powered directly from 3.3V, and each is independently
switched through the AP2502.