R1 case: finish most of the front side
- Holes cut and toleranced for ports and side buttons
- Added tolerancing on LCD module

All the major features are now present for the front half
of the case, although there's still nothing to affix the
two halves of the case together (the back still needs to
be modeled, but it will be a lot simpler than the front).

Thickness is now 18mm -- the LCD assembly takes up a lot
of space on the Z axis and the SWD port adds around 3mm.