WIP redesign with STM32H7R3Z8J6
The H7R3 has a bunch of features that greatly simplify the design:
support for HyperRAM, which requires far fewer pins than SDRAM and
is more power efficient; this frees up the FMC to interface with the
display in 8-bit 8080 MCU mode; and there is a built-in USB HS PHY,
removing the need for an external PHY.

The huge reduction in signal count allows for a smaller pin count
package -- only 144 pins. HyperRAM only comes in BGA packages so it
makes sense to also use the UFBGA-144 package for the CPU, which
takes up less than 1/4 of the board area of an LQFP-176 package.