R1-Rev2: convert buttons to SMD metal dome switches
Rubber domes increase mechanical complexity since they need to
be aligned and supported in the case to prevent them from going
off-center, and they require a lot of vertical space (~5mm).
Domes are also an additional component that has to be sourced
separately.

In contrast, prepackaged metal dome switches are normal SMD parts
and can be sourced and assembled alongside electronic components.
They are also much, much thinner and mechanical design is simpler.