R1-Rev2: use FMC NAND mode for 8-bit 8080 LCD interface
NAND mode will translate 16-bit AXI transfers into two
8-bit FMC transfers, which allows the 8-bit LCD bus to
work with DMA2D. The NWAIT pin does need to be "wasted"
on a fixed 3v3 input, since the FMC expects to poll it
(normally it would be the busy signal from a NAND chip).

It's seems likely this can be disabled by clearing the
PWAITEN bit so I've added a solder jumper to test it.

This saves 7 pins off the LCD data bus. The IM pins are
now fixed which saves another 2 pins.