R1-Rev2: change LCD ribbon cable slot placement and size
The LCD won't sit flat if the fat side of the cable is lying
underneath it because there are a bunch of capacitors there.
The cable is long enough to reach the back of the PCB though
if the entire thing is put through a slot right underneath
the screen, although it's annoyingly huge.

I've removed the screw holes there since the routing becomes
very tight in that area with them in place. Depending on how
it goes there might be room to bring one of them back or put
a screw in the middle, between the slot and the D-pad.