R1-Rev2: improve reset circuit, add XSPI CS pullup
Connect NRST to xspi_reset and the regulator power-good signals.
This should ensure that the STM32 and RAM will stay in reset
until both power rails have started up. It shouldn't be necessary
to control xspi_reset with a GPIO, but for now leave it connected
just in case.

The XSPI CS pin also requires an external pullup to guarantee it
will be pulled high during power-on, which is required to ensure
the RAM resets properly.