R1-Rev2: move LCD in Z-axis to free up board space
Mounting the LCD directly to the PCB just proved to be too
awkward; mainly because the connector position made it hard
to fit components on the upper half of the PCB. Mounting it
on top of a 1.2mm support which directly contacts the PCB
alleviates this, and the cable slot can now be smaller and
moved up closer to the connector. This provides a bit more
room everywhere and will make placement & routing simpler.

There is a slight impact on thickness, but the acrylic LCD
cover can be reduced in thickness to compensate.