R1-Rev2: move lcd.dcx pin to FMC_A19
DMA2D always auto-increments the destination address, unlike
normal DMA which has the option not to. The LCD is mapped as
an 8-bit memory and DMA2D will perform 16-bit transfers so
the number of pixels transferred must be <= 2**x / 2 if FMC_Ax
is the DCX pin; otherwise the transfer will end up toggling
DCX and won't operate correctly.

Since 320*240 pixels exceeds the limit if FMC_A16 is used as
the DCX pin, move DCX to a higher FMC address pin.